radiocovid-etl/pyproject.toml
[build-system]
build-backend = "uv_build"
requires = ["uv_build>=0.9.28,<0.10.0"]

[dependency-groups]
dev = [
    "jupyter>=1.0.0,<2.0.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "seaborn>=0.13.2",
]
lint = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
]
typing = ["mypy>=1.19.1"]
test = []

[project]
name = "radiocovid-etl"
version = "0.1.0"
requires-python = ">=3.10,<3.11"
dependencies = [
    "dvc>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "joblib>=1.5.3",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "hydra-core>=1.3.2",
    "tqdm>=4.67.1",
    "numpy>=1.26.4",
    "hydra-colorlog>=1.2.0",
    "opencv-python>=4.12.0.88",
    "openpyxl>=3.1.5",
    "scikit-image>=0.25.2",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.uv.build-backend]
module-name = "radiocovid.etl"

[tool.uv.sources]
radiocovid-core = { workspace = true }
